+++ lib/design_system/typography/app_text_style_constants.dart
@@ -1,0 +1,52 @@
+class AppTextStyleConstants {
+  static const double display1Size = 56;
+  static const double display2Size = 40;
+  static const double title1Size = 36;
+  static const double title2Size = 28;
+  static const double title3Size = 24;
+  static const double heading1Size = 22;
+  static const double heading2Size = 20;
+  static const double headline1Size = 18;
+  static const double headline2Size = 17;
+  static const double body1Size = 16;
+  static const double body2Size = 15;
+  static const double label1Size = 14;
+  static const double label2Size = 13;
+  static const double caption1Size = 12;
+  static const double caption2Size = 11;
+
+  static const double display1LetterSpacing = -3.19;
+  static const double display2LetterSpacing = -2.82;
+  static const double title1LetterSpacing = -2.70;
+  static const double title2LetterSpacing = -2.36;
+  static const double title3LetterSpacing = -2.30;
+  static const double heading1LetterSpacing = -1.94;
+  static const double heading2LetterSpacing = -1.20;
+  static const double headline1LetterSpacing = -0.02;
+  static const double headline2LetterSpacing = 0.0;
+  static const double body1LetterSpacing = 0.57;
+  static const double body2LetterSpacing = 0.96;
+  static const double label1LetterSpacing = 1.45;
+  static const double label2LetterSpacing = 1.94;
+  static const double caption1LetterSpacing = 2.52;
+  static const double caption2LetterSpacing = 3.11;
+
+  static const double display1LineHeight = 1.286;
+  static const double display2LineHeight = 1.3;
+  static const double title1LineHeight = 1.334;
+  static const double title2LineHeight = 1.358;
+  static const double title3LineHeight = 1.334;
+  static const double heading1LineHeight = 1.364;
+  static const double heading2LineHeight = 1.4;
+  static const double headline1LineHeight = 1.445;
+  static const double headline2LineHeight = 1.412;
+  static const double body1NormalLineHeight = 1.5;
+  static const double body1ReadingLineHeight = 1.625;
+  static const double body2NormalLineHeight = 1.467;
+  static const double body2ReadingLineHeight = 1.6;
+  static const double label1NormalLineHeight = 1.429;
+  static const double label1ReadingLineHeight = 1.571;
+  static const double label2LineHeight = 1.385;
+  static const double caption1LineHeight = 1.334;
+  static const double caption2LineHeight = 1.273;
+}
